+++ Client/includes/View/Creature.h
@@ -21,7 +21,7 @@ class Creature : public Unit {
    public:
     /* Constructor */
     Creature(Renderer* renderer, UnitSpriteContainer* sprites, const int tile_w,
-             const int tile_h, const int tile_movement_time);
+             const int tile_h, const float tile_movement_time);
 
     /* Deshabilitamos el constructor por copia. */
     Creature(const Creature&) = delete;
